Extract RingChart component
