+++ components/home-page.tsx
@@ -21,59 +21,7 @@ import * as d3 from "d3";
 import { JwtContext } from "@/app/lib/jwt-provider";
 import { useCurrentBudget } from "@/contexts/CurrentBudgetContext";
 import { useBudgets } from "@/contexts/BudgetContext";
-
-interface RingChartProps {
-  percentage: number;
-  color: string;
-  size?: number;
-}
-
-const RingChart: React.FC<RingChartProps> = ({
-  percentage,
-  color,
-  size = 100,
-}) => {
-  const strokeWidth = 10;
-  const radius = (size - strokeWidth) / 2;
-  const circumference = radius * 2 * Math.PI;
-  const strokeDashoffset = circumference - (percentage / 100) * circumference;
-
-  return (
-    <div className="relative">
-      <svg width={size} height={size} viewBox={`0 0 ${size} ${size}`}>
-        <circle
-          cx={size / 2}
-          cy={size / 2}
-          r={radius}
-          fill="none"
-          stroke="#e5e7eb"
-          strokeWidth={strokeWidth}
-        />
-        <circle
-          cx={size / 2}
-          cy={size / 2}
-          r={radius}
-          fill="none"
-          stroke={color}
-          strokeWidth={strokeWidth}
-          strokeDasharray={circumference}
-          strokeDashoffset={strokeDashoffset}
-          strokeLinecap="round"
-          transform={`rotate(-90 ${size / 2} ${size / 2})`}
-        />
-      </svg>
-      <div className="absolute inset-0 flex items-center justify-center">
-        <span className="text-2xl font-bold">{percentage}%</span>
-      </div>
-    </div>
-  );
-};
-
-interface Notification {
-  id: number;
-  message: string;
-  type: "warning" | "alert";
-}
+import { RingChart } from "./charts/RingChart";
 
 export function MainPage() {
   const [jwt, setAndStoreJwt] = useContext(JwtContext);
